Almost Finished Scene
My Scene Has almost all the features implemented into it assides for some polish in empty areas but the current tasks are still as follows:
Implement Matt's Water Section.
Implement or improve the Moving pilar  area near spawn.
Enemies Will now chase you in the upper cavern areas located near the yellow crystals
A Script which will lock off the finishing area / Tower until all the crystals have been placed in the big storage area (Just before the caverns) Note: this could be achieved by turning the path objects on and off based on if the player has passed global flags in the world. (and for simplicities sake the path can be loaded in chunks)
when the goal has been reached the game will end.